style: rename `sliceutils` package to `sliceutil`
